tailcfg: move views into tailcfg_view.go
